extracted some components and removed extra code
